requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
clarifai-grpc==8.8.0
colorama==0.4.6
emoji==2.1.0
exceptiongroup==1.0.1
flake8==5.0.4
frozenlist==1.3.1
googleapis-common-protos==1.56.4
greenlet==2.0.0.post0
grpcio==1.49.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
psycopg2==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-sqlalchemy-mock==0.1.4
python-telegram-bot==20.0a4
python-telegram-bot-raw==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
tomli==2.0.1
tornado==6.2
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1
